requirements.txt
mcp>=1.0.0
mysql-connector-python>=9.1.0
requests